+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+- Adds `retrieveEstimatedDeliveryDate` function to the Shipment class
+
 ## v6.5.0 (2023-04-18)
 
 - Adds beta `retrieveCarrierMetadata` function